Make SearchBar component more generic so it can be used for a variety of modules
