+++ src/js/jsx/search/SearchBar.jsx
@@ -31,11 +31,7 @@ define(function (require, exports, module) {
         Datalist = require("jsx!js/jsx/shared/Datalist"),
         Immutable = require("immutable");
 
-    var layerLib = require("adapter/lib/layer"),
-        pathUtil = require("js/util/path"),
-        svgUtil = require("js/util/svg"),
-        mathUtil = require("js/util/math"),
-        collection = require("js/util/collection"),
+    var svgUtil = require("js/util/svg"),
         strings = require("i18n!nls/strings");
     
     /**
@@ -52,19 +48,28 @@ define(function (require, exports, module) {
      * @const
      * @type {object} 
     */
-    var HEADERS = strings.SEARCH.HEADERS,
-        CATEGORIES = strings.SEARCH.CATEGORIES;
+    var CATEGORIES = strings.SEARCH.CATEGORIES;
 
     var SearchBar = React.createClass({
         mixins: [FluxMixin],
 
         propTypes: {
-            dismissDialog: React.PropTypes.func
+            dismissDialog: React.PropTypes.func,
+            // Function to get all possible options
+            getOptions: React.PropTypes.isRequired,
+            // Function to perform an action when an option is confirmed
+            executeOption: React.PropTypes.func.isRequired,
+            // Broad categories of what SearchBar has as options
+            searchTypes: React.PropTypes.arrayOf(React.PropTypes.string).isRequired,
+            // List of more specific categories that correlate with searchTypes
+            // Indicate that there are no categories for a search type with null
+            searchCategories: React.PropTypes.arrayOf(Immutable.List)
         },
 
         getDefaultProps: function () {
             return {
-                dismissDialog: _.identity
+                dismissDialog: _.identity,
+                searchCategories: null
             };
         },
 
@@ -87,7 +92,7 @@ define(function (require, exports, module) {
         },
 
         /**
-         * Select the layer or document specified and dismiss the dialog
+         * Perform action based on ID
          *
          * @param {string} id
          */
@@ -97,40 +102,11 @@ define(function (require, exports, module) {
                 return;
             }
 
-            // ID has type as first word, followed by the layer/document ID
-            var idArray = id.split("_"),
-                type = idArray[0],
-                idInt = mathUtil.parseNumber(idArray[1]),
-                flux = this.getFlux();
-
-            switch (type) {
-            case "filter":
+            if (id.indexOf("filter") === 0) {
                 this._updateFilter(id);
-                return;
-            case "layer":
-                var document = flux.store("application").getCurrentDocument(),
-                    selected = document.layers.byID(idInt);
-
-                if (selected) {
-                    flux.actions.layers.select(document, selected);
-                }
-                break;
-            case "curr-doc":
-                var selectedDoc = flux.store("document").getDocument(idInt);
-                
-                if (selectedDoc) {
-                    flux.actions.documents.selectDocument(selectedDoc);
-                }
-                break;
-            case "recent-doc":
-                var appStore = this.getFlux().store("application"),
-                    recentFiles = appStore.getRecentFiles(),
-                    fileName = recentFiles.get(idInt);
-
-                flux.actions.documents.open(fileName);
-                break;
+            } else {
+                this.props.executeOption(id);
             }
-            this.props.dismissDialog();
         },
 
         /**
@@ -143,7 +119,7 @@ define(function (require, exports, module) {
                 filterValues = _.drop(idArray),
                 updatedFilter = id ? _.uniq(this.state.filter.concat(filterValues)) : [],
                 filterIcons = svgUtil.getSVGClassesFromLayerTypes(updatedFilter);
-                
+
             this.setState({
                 filter: updatedFilter,
                 icons: filterIcons
@@ -151,214 +127,52 @@ define(function (require, exports, module) {
 
             this.refs.datalist.resetInput(idArray, filterIcons.length);
         },
-
+        
         /**
-         * Get the layer ancestry
-         *
-         * @private
-         * @param {Layer} layer
-         * @return {string}
-         */
-        _formatLayerAncestry: function (layer) {
-            var layerTree = this.getFlux().store("application").getCurrentDocument().layers,
-                ancestors = layerTree.ancestors(layer),
-                ancestorNames = ancestors.first().name;
-                
-            return ancestors.skip(1).reduce(function (ancestors, ancestor) {
-                return ancestorNames += ("/" + ancestor.name);
-            }, ancestorNames);
-        },
-
-        /**
-         * Get the layer type and if it is linked or an artboard, as an array of strings.
-         * This is used for comparison against the inputted search terms and any search filters.
-         * Not directly user-visible, but needs to be able to match user input.
-         *
-         * @private
-         * @param {Layer} layer
-         * @return {Array.<string>}
-        */
-        _getLayerCategory: function (layer) {
-            var layerType = [CATEGORIES.LAYER];
-
-            if (layer.kind === layer.layerKinds.GROUP && layer.isArtboard) {
-                layerType.push(CATEGORIES.ARTBOARD);
-            } else {
-                // Find string associated with layer.kind, which is a number
-                _.forEach(Object.keys(layer.layerKinds), function (type) {
-                    if (layer.kind === layer.layerKinds[type]) {
-                        layerType.push(CATEGORIES[type].replace(" ", ""));
-                    }
-                });
-            }
-
-            if (layer.isLinked) {
-                layerType.push(CATEGORIES.LINKED);
-            }
-
-            return layerType;
-        },
-
-        /**
-         * Make list of layers in the current document to be used as dropdown options
+         * Make list of search category dropdown options
          * 
          * @return {Array.<object>}
          */
-        _getLayerOptions: function () {
-            // Get list of layers
-            var appStore = this.getFlux().store("application"),
-                document = appStore.getCurrentDocument(),
-                layers = document.layers.allVisibleReversed,
-                layerMap = layers.map(function (layer) {
-                    var ancestry = this._formatLayerAncestry(layer),
-                        layerType = this._getLayerCategory(layer),
-                        iconID = svgUtil.getSVGClassFromLayer(layer);
+        _getFilterOptions: function () {
+            var allFilters;
 
-                    return {
-                        id: "layer_" + layer.id.toString(),
-                        title: layer.name,
-                        info: ancestry,
-                        displayInfo: ancestry,
-                        svgType: iconID,
-                        category: layerType,
-                        type: "item"
-                    };
-                }.bind(this)),
+            this.props.searchTypes.forEach(function (types, header, index) {
+                var allCategories = this.props.searchCategories,
+                    filters = Immutable.List();
+                if (allCategories) {
+                    var categories = allCategories[index];
+                                       
+                    filters = categories ? categories.map(function (kind) {
+                        var idType = CATEGORIES[kind],
+                            title = CATEGORIES[kind];
 
-                // Get shortest unique layer ancestry
-                ancestors = collection.pluck(layerMap, "info"),
-                shortenedPaths = pathUtil.getShortestUniquePaths(ancestors).toJS();
+                        title = title.charAt(0).toUpperCase() + title.slice(1);
+                        idType = idType.replace(" ", "");
 
-            layerMap.forEach(function (layerItem, index) {
-                var newInfo = shortenedPaths[index];
-                
-                // Only show ancestry if other layers have the same name
-                if (layerItem.title === newInfo) {
-                    layerItem.displayInfo = "";
-                } else {
-                    layerItem.displayInfo = newInfo;
+                        return {
+                            id: "filter_" + CATEGORIES[header] + "_" + idType,
+                            title: title,
+                            category: [CATEGORIES[header], idType],
+                            type: "item"
+                        };
+                    }) : filters;
                 }
-            });
 
-            var layerLabel = {
-                id: "layer_header",
-                title: HEADERS.LAYERS,
-                type: "header"
-            },
-            layerOptions = layerMap.unshift(layerLabel);
-
-            return layerOptions;
-        },
-
-        /**
-         * Make list of currently open documents to be used as dropdown options
-         * 
-         * @return {Array.<object>}
-         */
-        _getCurrDocOptions: function () {
-            var appStore = this.getFlux().store("application"),
-                docStore = this.getFlux().store("document"),
-                document = appStore.getCurrentDocument(),
-                openDocs = appStore.getOpenDocumentIDs().filterNot(function (doc) {
-                                return doc === document.id;
-                            }),
-                docMap = openDocs.map(function (doc) {
-                    return {
-                        id: "curr-doc_" + doc.toString(),
-                        title: docStore.getDocument(doc).name,
-                        type: "item",
-                        category: [CATEGORIES.DOCUMENT, CATEGORIES.CURRENT]
-                    };
-                }),
-                docLabel = {
-                    id: "curr-doc_header",
-                    title: HEADERS.CURRENT_DOCS,
-                    type: "header"
-                },
-
-                docOptions = docMap.unshift(docLabel);
-
-            return docOptions;
-        },
-
-        /**
-         * Make list of recent documents to be used as dropdown options
-         * 
-         * @return {Array.<object>}
-         */
-        _getRecentDocOptions: function () {
-            var appStore = this.getFlux().store("application"),
-                recentFiles = appStore.getRecentFiles(),
-                recentDocMap = recentFiles.map(function (doc, index) {
-                    return {
-                        id: "recent-doc_" + index.toString(),
-                        title: pathUtil.getShortestUniquePaths(Immutable.List.of(doc)).toJS()[0],
-                        type: "item",
-                        info: doc,
-                        displayInfo: doc,
-                        category: [CATEGORIES.DOCUMENT, CATEGORIES.RECENT]
-                    };
-                });
-            
-            // Get shortest unique file path
-            var paths = collection.pluck(recentDocMap, "info"),
-                shortenedPaths = pathUtil.getShortestUniquePaths(paths).toJS();
-
-            recentDocMap.forEach(function (docItem, index) {
-                docItem.displayInfo = shortenedPaths[index];
-            });
-
-            var recentDocLabel = {
-                id: "recent-doc_header",
-                title: HEADERS.RECENT_DOCS,
-                type: "header"
-            };
-
-            return recentDocMap.unshift(recentDocLabel);
-        },
-
-        /**
-         * Make list of search category dropdown options based on header
-         * 
-         * @param {string} header, either "LAYER" or "DOCUMENT"
-         * @return {Array.<object>}
-         */
-        _getFilterOptions: function (header) {
-            var categoryList = header === "DOCUMENT" ? ["CURRENT", "RECENT"] : Object.keys(layerLib.layerKinds),
-                categories = Immutable.List(categoryList).filterNot(function (kind) {
-                    return (kind === "ANY" || kind === "GROUPEND" ||
-                        kind === "3D" || kind === "VIDEO");
-                });
-                                   
-            var filters = categories.map(function (kind) {
-                var idType = CATEGORIES[kind],
-                    title = CATEGORIES[kind];
-
-                title = title.charAt(0).toUpperCase() + title.slice(1);
-                idType = idType.replace(" ", "");
-
-                return {
-                    id: "filter_" + CATEGORIES[header] + "_" + idType,
-                    title: title,
-                    category: [CATEGORIES[header], idType],
+                // To search for all layers, etc
+                var headerTitle = CATEGORIES[header];
+                headerTitle = headerTitle.charAt(0).toUpperCase() + headerTitle.slice(1);
+                
+                var headerFilter = {
+                    id: "filter_" + CATEGORIES[header],
+                    title: headerTitle,
+                    category: [CATEGORIES[header]],
                     type: "item"
                 };
-            });
+                filters = filters.unshift(headerFilter);
+                allFilters = typeof (allFilters) === "undefined" ? filters : allFilters.concat(filters);
+            }.bind(this, allFilters));
 
-            // To search for all layers, documents, etc
-            var headerTitle = CATEGORIES[header];
-            headerTitle = headerTitle.charAt(0).toUpperCase() + headerTitle.slice(1);
-            
-            var headerFilter = {
-                id: "filter_" + CATEGORIES[header],
-                title: headerTitle,
-                category: [CATEGORIES[header]],
-                type: "item"
-            };
-            
-            filters = filters.unshift(headerFilter);
-
-            return filters;
+            return Immutable.List(allFilters);
         },
 
         /**
@@ -366,11 +180,10 @@ define(function (require, exports, module) {
          * @return {Array.<object>}
          */
         _getAllSelectOptions: function () {
-            var filterOptions = this._getFilterOptions("LAYER").concat(this._getFilterOptions("DOCUMENT")),
-                layerOptions = this._getLayerOptions(),
-                docOptions = this._getCurrDocOptions().concat(this._getRecentDocOptions());
-           
-            return filterOptions.concat(layerOptions).concat(docOptions);
+            var filterOptions = this._getFilterOptions(),
+                options = this.props.getOptions();
+            
+            return filterOptions.concat(options);
         },
 
         /**
@@ -381,12 +194,7 @@ define(function (require, exports, module) {
          * @return {Array.<string>}
          */
         _getFilterIcons: function (filter) {
-            // currently only have icons for layers
-            if (filter.length > 1 && filter.join(" ").indexOf("layer") > -1) {
-                return svgUtil.getSVGClassesFromLayerTypes(filter);
-            } else {
-                return ["tool-rectangle"]; // standin for non-layers
-            }
+            return svgUtil.getSVGClassesFromLayerTypes(filter);
         },
 
         /**
@@ -447,7 +255,7 @@ define(function (require, exports, module) {
 
                 // If option has info, search for it with and without '/' characters
                 // Don't check each word of search term individually because want 
-                // search to preserve order of layer hierarchy
+                // search to preserve order of layer hierarchy or file path
                 var info = option.displayInfo ? option.displayInfo.toLowerCase() : "",
                     searchableInfo = info.concat(info.replace(/\//g, " "));
