+++ src/nls/root/strings.js
@@ -576,9 +576,9 @@ define(function (require, exports, module) {
         SEARCH: {
             PLACEHOLDER: "Type to search",
             HEADERS: {
-                LAYERS: "Layers",
-                CURRENT_DOCS: "Current Documents",
-                RECENT_DOCS: "Recent Documents"
+                LAYER: "Layers",
+                CURRENT_DOC: "Current Documents",
+                RECENT_DOC: "Recent Documents"
             },
             CATEGORIES: {
                 DOCUMENT: "document",
